Program 5 Adrian: Incomplete Read Below
A lot of the added functionality comes in the form of deciding what to do based on what input is given i.e. if User loads and game state has changed prompt for "Save?" then prompt for "Load Default" or "Other" then prompt for "Load from save files?" or "Load from Path" i.e. load a completely new maze.  Also adjusting stuff like if I haven't change the board state Load should NOT prompt for a load default again but rather should prompt for a "Load new Game Using Default?"  and if that is NO then "Load a new maze from the Library" i.e. load a Save and if that is NO then Load from a Path i.e a completely new Maze